refactor: move socket method to route/
