+++ controllers/notificationController.js
@@ -1,6 +1,6 @@
 const notificationModel	= require('../models/notificationModel');
 const token				= require('../models/token');
-const MESSAGE_TYPE 		= require('../utilities/socket-message').MESSAGE_TYPE;
+const MESSAGE_TYPE 		= require('../routes/socket-message').MESSAGE_TYPE;
 
 const getExpiredProjectId = async (req, res) => {
 	let userToken;
